+++ common/src/main/kotlin/com/cobblemon/mod/common/client/gui/PokemonGuiUtils.kt
@@ -56,6 +56,7 @@ fun drawProfilePokemon(
 
     if (state != null) {
         model.getPose(PoseType.PROFILE)?.let { state.setPose(it.poseName) }
+        state.timeEnteredPose = 0F
         model.setupAnimStateful(null, state, 0F, 0F, 0F, 0F, 0F)
     } else {
         model.setupAnimStateless(PoseType.PROFILE)