+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/PoseableEntityState.kt
@@ -32,7 +32,8 @@ abstract class PoseableEntityState<T : Entity> {
     val statefulAnimations: MutableList<StatefulAnimation<T, *>> = mutableListOf()
     val quirks = mutableMapOf<ModelQuirk<T, *>, QuirkData<T>>()
     val additives: MutableList<PosedAdditiveAnimation<T>> = mutableListOf()
-    var animationPreviousSeconds = 0F
+    var timeEnteredPose = 0F
+    var previousAnimationSeconds = 0F
     var animationSeconds = 0F
     var deltaSeconds = 0F
     var timeLastRendered = System.currentTimeMillis()
@@ -62,7 +63,7 @@ abstract class PoseableEntityState<T : Entity> {
 
         timeLastRendered = now
         deltaSeconds = deltaMillis / 1000F
-        animationPreviousSeconds = animationSeconds
+        previousAnimationSeconds = animationSeconds
         animationSeconds += deltaSeconds
     }
 
@@ -76,8 +77,8 @@ abstract class PoseableEntityState<T : Entity> {
         if (model != null) {
             val poseImpl = model.getPose(pose) ?: return
             poseImpl.onTransitionedInto(this)
+            timeEnteredPose = animationSeconds
         }
-
     }
 
     fun applyAdditives(entity: T?, model: PoseableEntityModel<T>, state: PoseableEntityState<T>?) {